backend/requirements.txt
-i https://pypi.org/simple

aiopyston==1.2.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
datasketch==1.6.5
dependency-injector==4.48.1
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
injector==0.22.0
iniconfig==2.1.0
itsdangerous==2.2.0
javalang==0.13.0
jinja2==3.1.6
lxml==6.0.0
markupsafe==3.0.2
more-itertools==10.7.0
mosspy==1.0.9
openpyxl==3.1.5
packaging==25.0
pam==0.2.0
paramiko==4.0.0
pluggy==1.6.0
py==1.11.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.1
pytest==8.3.3
pytest-mock==3.14.1
python-pam==2.0.2
pyyaml==6.0.2
requests==2.32.4
scikit-learn==1.4.2
scp==0.15.0
sentry-sdk[flask]==2.35.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.43
tap.py[yaml]==3.2.1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3

pyjwt==2.9.0